Added security using JWT
Implemented Only teacher who create assignment can edit and delete it

+++ resources/assignment.py
@@ -3,7 +3,9 @@ from flask_smorest import Blueprint, abort
 from schemas import AssignmentSchema,AssignmentUpdateSchema
 from db import db
 from models.assignmentmodel import AssignmentModel
+from flask_jwt_extended import jwt_required, get_jwt,current_user
 from sqlalchemy.exc import SQLAlchemyError,IntegrityError
+
 
 blp = Blueprint("Assignment",__name__,description="Assignment service")
 
@@ -17,9 +19,19 @@ class Assignment(MethodView):
         assignment = AssignmentModel.query.get_or_404(assignment_id,description=f"No assignment found")
         return assignment
     
-    # DELETE method for DELETING specific assignment using Assignment ID 
+    # DELETE method for DELETING specific assignment using Assignment ID
+    @jwt_required() 
     def delete(self,assignment_id):
+        
+    
+        if(current_user.usertype!="teacher"):
+            abort(401, message="Teacher privilege required.")
+
         assignment = AssignmentModel.query.get_or_404(assignment_id)
+        
+        ## Checking current user id from JWT with teacher id who created assignment
+        if(assignment.teacher_id!=current_user.id):
+            abort(401,message="Only teacher which created assignment can delete it")
         try:
             db.session.delete(assignment)
             db.session.commit()
@@ -28,18 +40,30 @@ class Assignment(MethodView):
         return {"message":"Assignment deleted"}, 200
     
     # PUT method for UPDATING specific assignment using Assignment ID 
+    @jwt_required()
     @blp.arguments(AssignmentUpdateSchema)
     @blp.response(200,AssignmentUpdateSchema)
     def put(self,assignment_data,assignment_id):
+
+        if(current_user.usertype!="teacher"):
+            abort(401, message="Teacher privilege required.")
+
         update_title = assignment_data.get("title")
         update_desc = assignment_data.get("desc")
-        
+        update_teacher = assignment_data.get("teacher_id")
+
         assignment = AssignmentModel.query.get_or_404(assignment_id,description=f"No assignment found")
+
+        ## Checking current user id from JWT with teacher id who created assignment
+        if(assignment.teacher_id!=current_user.id):
+            abort(401,message="Only teacher which created assignment can delete it")
         
         if update_title is not None:
             assignment.title = update_title
         if update_desc is not None:
             assignment.desc = update_desc
+        if update_teacher is not None:
+            assignment.teacher_id = update_teacher
         try:
             db.session.add(assignment)
             db.session.commit()
@@ -49,6 +73,7 @@ class Assignment(MethodView):
 
         return assignment
     
+
 
 # ASSIGNMENT LIST CLASS for fetching list of assignments and adding a new assignment
 @blp.route("/assignment")
@@ -60,18 +85,30 @@ class AssignmentList(MethodView):
         return AssignmentModel.query.all()
 
     # POST method to add new assignment
+    @jwt_required()
     @blp.arguments(AssignmentSchema)
     @blp.response(201,AssignmentSchema)
     def post(self,assignment_data):
+
+        # Allowing only teacher to create assignment using jwt claims
+        # jwt = get_jwt()
+        # if not jwt.get("is_teacher"):
+        #     abort(401, message="Teacher privilege required.")
+        
+        ## Checking current user id from JWT with teacher id who created assignment
+        if(current_user.usertype!="teacher"):
+            abort(401, message="Teacher privilege required.")
+
         title = assignment_data.get("title")
         desc = assignment_data.get("desc")
+        teacher_id = current_user.id
 
-        assignment = AssignmentModel(title=title,desc=desc)
+        assignment = AssignmentModel(title=title,desc=desc,teacher_id=teacher_id)
         try:
             db.session.add(assignment)
             db.session.commit()
         except IntegrityError:
-            abort(400,message="An Assignment with that title already exists.")
+            abort(409,message="An Assignment with that title already exists or Teacher with that id do not exists")
         except SQLAlchemyError as e:
             abort(500,message="An error occured while creating assignment.")
